backend/requirements.txt
aiohttp==3.8.1
pydantic==1.9.1
asyncpg==0.26.0
tortoise-orm==0.19.2
aerich==0.6.3
